auto-optimizer/requirements.txt
onnx >= 1.8.0
numpy >= 1.21.0
onnx-simplifier >= 0.3.6
protobuf >= 3.7.0
pillow >= 9.0.0
onnxruntime >= 1.10.0
setuptools >= 63.2.0
tqdm >= 4.64.0
pyacl @ git+https://gitee.com/peng-ao/pyacl.git